boost/requirements.txt
amqp==2.5.0
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
Babel==2.7.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
celery==4.3.0
cffi==1.12.3
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.2.3
djangorestframework==3.10.1
elasticsearch==7.0.2
flower==0.9.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
kombu==4.6.3
lxml==4.3.4
parsel==1.5.1
Pillow==6.1.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dotenv==0.10.3
pytz==2019.1
queuelib==1.5.0
Scrapy==1.7.1
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tornado==5.1.1
Twisted==19.2.1
urllib3==1.25.3
vine==1.3.0
w3lib==1.20.0
zope.interface==4.6.0
